+++ tests/Debug/RouterCollectorTest.php
@@ -35,7 +35,7 @@ final class RouterCollectorTest extends AbstractCollectorTestCase
     public function testWithoutCurrentRoute(): void
     {
         $collector = new RouterCollector(
-            new SimpleContainer()
+            new SimpleContainer(),
         );
         $collector->startup();
 
@@ -55,7 +55,7 @@ final class RouterCollectorTest extends AbstractCollectorTestCase
         $collector = new RouterCollector(
             new SimpleContainer([
                 CurrentRoute::class => $currentRoute,
-            ])
+            ]),
         );
         $collector->startup();
 
@@ -64,8 +64,21 @@ final class RouterCollectorTest extends AbstractCollectorTestCase
         $this->assertSame(['currentRoute'], array_keys($collected));
         $this->assertSame(
             ['matchTime', 'name', 'pattern', 'arguments', 'hosts', 'uri', 'action', 'middlewares'],
-            array_keys($collected['currentRoute'])
+            array_keys($collected['currentRoute']),
         );
+    }
+
+    public function testWithCurrentRouteWithoutMatchedRoute(): void
+    {
+        $collector = new RouterCollector(
+            new SimpleContainer([
+                CurrentRoute::class => new CurrentRoute(),
+            ]),
+        );
+        $collector->startup();
+
+        $this->assertSame([], $collector->getSummary());
+        $this->assertSame(['currentRoute' => null], $collector->getCollected());
     }
 
     /**
@@ -105,15 +118,15 @@ final class RouterCollectorTest extends AbstractCollectorTestCase
         $this->assertArrayHasKey('routeTime', $data);
         $this->assertEquals(
             $this->container->get(RouteCollectionInterface::class)->getRoutes(),
-            $data['routes']
+            $data['routes'],
         );
         $this->assertEquals(
             $this->container->get(RouteCollectionInterface::class)->getRouteTree(),
-            $data['routesTree']
+            $data['routesTree'],
         );
         $this->assertEquals(
             0.001,
-            $data['routeTime']
+            $data['routeTime'],
         );
     }