+++ tests/Support/NotFoundException.php
@@ -7,6 +7,4 @@ namespace Yiisoft\Router\Tests\Support;
 use Psr\Container\NotFoundExceptionInterface;
 use RuntimeException;
 
-final class NotFoundException extends RuntimeException implements NotFoundExceptionInterface
-{
-}
+final class NotFoundException extends RuntimeException implements NotFoundExceptionInterface {}