+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 4.0.3 under development
 
-- Enh #276: Explicitly import classes, functions, and constants in the "use" section (@rustamwin)
+- Enh #276, #292: Explicitly import classes, functions, and constants in the "use" section (@rustamwin, @vjik)
 - Enh #277, #281: Remove restrictions from `prependMiddleware()` and `middleware()` methods (@klsoft-web, @vjik)
 
 ## 4.0.2 December 13, 2025